+++ 1/adapters/__init__.py
@@ -1,0 +1,9 @@
+# This file makes the adapters directory a Python package
+# Import adapter classes
+from .BaseAdapter import BaseAdapter
+from .BashAdapter import BashAdapter
+from .PHPAdapter import PHPAdapter
+from .NodeAdapter import NodeAdapter
+from .HttpClientAdapter import HttpClientAdapter
+from .HttpServerAdapter import HttpServerAdapter
+from .PythonAdapter import PythonAdapter